Création des entités order + création premiere étape du tunnel d'achat avant de rentrer la commande en bdd

+++ src/Classe/Cart.php
@@ -9,12 +9,12 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
 class Cart{
 
     private $session;
-    private $entityManagaer;
+    private $entityManager;
 
     public function __construct(SessionInterface $session, EntityManagerInterface $entityManager)
     {
         $this->session = $session;
-        $this->entityManagaer = $entityManager;
+        $this->entityManager = $entityManager;
     }
 
     public function add ($id){
@@ -58,23 +58,27 @@ class Cart{
         return $this->session->set('cart',$cart);
     }
 
-    public function getFull(){
-
-
+    public function getFull()
+    {
         $cartComplete = [];
 
-        foreach($this->get() as $id => $quantity){
-            $product_object = $this->entityManagaer->getRepository(Product::class)->findOneBy(['id' => $id]);
-            if(!$product_object){
-                $this->delete($id);
-                continue;
+        if ($this->get()) {
+            foreach ($this->get() as $id => $quantity) {
+                $product_object = $this->entityManager->getRepository(Product::class)->findOneById(['id' => $id]);
+
+                if (!$product_object) {
+                    $this->delete($id);
+                    continue;
+                }
+
+                $cartComplete[] = [
+                    'product' => $product_object,
+                    'quantity' => $quantity
+                ];
             }
-            $cartComplete[] = [
-                'product' => $product_object,
-                'quantity' => $quantity
-            ];
         }
 
         return $cartComplete;
     }
+
 }
